pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-purple-ops"
version = "1.0.0"
description = "Vendor-neutral harness for AI safety, security, and compliance evaluation"
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [{ name = "AI Purple Ops Authors" }]
dependencies = [
    "pyyaml>=6.0.2,<7",
    "python-dotenv>=1.0.1,<2",
    "rich>=13.7.1,<14",
    "tenacity>=8.3.0,<9",
    "typer>=0.12.5,<1",
    "jsonschema>=4.23.0,<5",
    "junit-xml>=1.9.0,<2",
    "requests>=2.31.0,<3",  # For adapter HTTP connections
    "pygad>=3.3.0,<4",  # Genetic algorithm library
    "duckdb>=1.0.0,<2",  # Analytics database for mutation tracking
    "pyrit>=0.4.0,<1",  # For DuckDB schema and memory management
    "torch>=2.1.0,<3",  # Required for GCG white-box mode
    "tqdm>=4.66.0,<5",  # Progress bars for GCG iterations
    "scipy>=1.9.0,<2",  # Statistical functions for confidence intervals
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "pytest-cov>=5.0,<6",
  "hypothesis>=6.100,<7",
  "ruff>=0.6,<1",
  "black>=24.8,<25",
  "mypy>=1.11,<2",
  "bandit>=1.7.9,<2",
  "pip-audit>=2.7,<3",
  "pre-commit>=3.8,<4",
  "types-PyYAML>=6.0.12,<7",
  "types-requests>=2.31,<3",
]
local = [
    "transformers>=4.36.0",  # For tokenizer access in GCG
    "accelerate>=0.25.0",  # For GPU acceleration
    "bitsandbytes>=0.41.0",  # For quantization
]
adversarial = [
    "transformers>=4.36.0",  # For tokenizer access
    "accelerate>=0.25.0",  # For GPU acceleration
    "nanogcg>=0.1.0",  # True gradient-based GCG optimization
]
# Official attack implementations (for direct import, not subprocess)
# Note: Most users should use plugin system (aipop plugins install) instead
pair-official = [
    "fschat==0.2.23",  # FastChat for model serving
    "anthropic>=0.8.0",  # Claude API
    "google-generativeai>=0.3.0",  # Gemini API
    "wandb>=0.16.0",  # Experiment tracking
    "pandas>=2.0.0",  # Data processing
]
autodan-official = [
    "fschat==0.2.20",  # FastChat (different version from PAIR)
    "transformers==4.28.0",  # Specific version required by AutoDAN
    "sentencepiece>=0.1.99",  # Tokenization
    "nltk>=3.8.1",  # Text processing
]
# GCG has no PyPI package - install via plugin system only
all-official = [
    "ai-purple-ops[pair-official,autodan-official]",
]
# Legacy implementations (no extra deps, for air-gapped environments)
legacy = []
llamacpp = [
  "llama-cpp-python>=0.2.0",
]
cloud = [
  "openai>=1.6.0",
  "anthropic>=0.8.0",
  "boto3>=1.34.0",  # For AWS Bedrock
]
all-adapters = [
  "ai-purple-ops[local,llamacpp,cloud]",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
addopts = "-q --strict-markers"
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "requires_gpu: marks tests that require GPU (deselect with '-m \"not requires_gpu\"')",
    "requires_official: marks tests that require official plugin repos installed (deselect with '-m \"not requires_official\"')",
    "requires_api: marks tests that require API keys (deselect with '-m \"not requires_api\"')",
    "integration: marks integration tests (deselect with '-m \"not integration\"')",
]

[tool.setuptools]
packages = { find = { where = ["src", "."] } }

[project.scripts]
aipop = "cli.harness:main"
